+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # argusai
 
+## 0.12.0
+
+### Patch Changes
+
+- Updated dependencies
+  - argusai-core@0.12.0
+  - argusai-dashboard@0.12.0
+  - argusai-mcp@0.12.0
+
 ## 0.11.0
 
 ### Patch Changes