+++ packages/dashboard/CHANGELOG.md
@@ -1,5 +1,12 @@
 # argusai-dashboard
 
+## 0.12.0
+
+### Patch Changes
+
+- Updated dependencies
+  - argusai-core@0.12.0
+
 ## 0.11.0
 
 ### Patch Changes