+++ packages/server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # argusai-server
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies
+  - argusai-core@0.12.0
+
 ## 0.6.2
 
 ### Patch Changes